agent/requirements.txt
livekit>=0.9.0
livekit-agents>=0.4.0
livekit-plugins-openai>=0.2.0
livekit-plugins-silero>=0.2.0
groq>=0.1.0
fpdf>=1.7.2
PyPDF2>=1.26.0
langchain
langchain-community
llama-parse
fastembed
chromadb
python-dotenv
langchain-groq
chainlit
fastembed
unstructured[md]
ollama
langchain_community